Updated the way render() works.
The render() method in \Jolt\Controller only takes one parameter
now: the name of the view to render. Leave empty if you want
to use the name of the method.

Two new methods: renderView() and renderToBlock() were added.
The renderView() method renders a view and returns it. The
renderToBlock() method renders a view and stores it in a block.

render() was updated to use renderView() as well.